Move surface_texture.h to common/
