+++ books/RayTracingTheNextWeek.html
@@ -838,6 +838,8 @@ classes so feel free to do something different, but I am a big believer in this 
 being able to make any color a texture is great.
 
     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ C++
+    #include "common/rtweekend.h"
+
     class texture {
         public:
             virtual vec3 value(double u, double v, const vec3& p) const = O;
@@ -1071,6 +1073,8 @@ Now if we create an actual texture that takes these floats between 0 and 1 and c
 colors:
 
     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ C++
+    #include "common/perlin.h"
+
     class noise_texture : public texture {
         public:
             noise_texture() {}
@@ -1492,6 +1496,8 @@ utility `stb_image`. It reads in an image into a big array of unsigned char. The
 RGBs that each range 0..255 for black to fully-on.
 
     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ C++
+    #include "common/texture.h"
+
     class image_texture : public texture {
         public:
             image_texture() {}
