+++ src/common/surface_texture.h
